Move IRepository & ApiController into Shared
