+++ Fitfuel.Server/Fitfuel.sln
@@ -195,5 +195,7 @@ Global
 		{0707B918-21F8-4ABF-AD93-296704D6671D} = {F6D3C2C6-E561-404D-9E87-690D7A69A622}
 		{AD66C549-4C8D-4048-8EC3-2F484B7CBC40} = {F6D3C2C6-E561-404D-9E87-690D7A69A622}
 		{229C685F-98F0-449F-BA4D-DC4F23378740} = {F6D3C2C6-E561-404D-9E87-690D7A69A622}
+		{8A00D1D6-A858-4C01-83C0-75BDC7AE3CC6} = {58BC99B7-1EA1-4712-AC93-3499A6814EE3}
+		{F34BABC7-150F-4A78-BB05-F6D8DE127F26} = {9EE9A9C0-E89C-4F79-99B0-360B71C2BD1F}
 	EndGlobalSection
 EndGlobal
