+++ Fitfuel.Server/Modules/Workouts/Fitfuel.Workouts.API/Controllers/ExerciseController.cs
@@ -1,4 +1,5 @@
-﻿using Fitfuel.Workouts.Application.Abstractions;
+﻿using Fitfuel.Shared.Presentation;
+using Fitfuel.Workouts.Application.Abstractions;
 using FitFuel.Workouts.Contracts.Exercises;
 using MapsterMapper;
