+++ Fitfuel.Server/Modules/Workouts/Fitfuel.Workouts.API/Fitfuel.Workouts.API.csproj
@@ -13,4 +13,8 @@
       <ProjectReference Include="..\Fitfuel.Workouts.Shared\Fitfuel.Workouts.Shared.csproj" />
     </ItemGroup>
 
+    <ItemGroup>
+      <Folder Include="Common\" />
+    </ItemGroup>
+
 </Project>
