+++ Fitfuel.Server/Modules/Workouts/Fitfuel.Workouts.Application/Services/WorkoutPlanService.cs
@@ -1,7 +1,7 @@
 ﻿using Fitfuel.Workouts.Domain.WorkoutPlanAggregate;
 using ErrorOr;
+using Fitfuel.Shared.Infrastructure.Abstractions;
 using Fitfuel.Workouts.Application.Abstractions;
-using Fitfuel.Workouts.Application.Abstractions.Persistence;
 using Fitfuel.Workouts.Application.Specifications.WorkoutPlans;
 using FitFuel.Workouts.Contracts.WorkoutPlans;
 using Fitfuel.Workouts.Domain.Common.Errors;
